requirements.txt
asgiref==3.3.4
Django==3.2
django-bootstrap3==15.0.0
django-debug-toolbar==3.2.2
django-environ==0.4.5
gunicorn==20.1.0
importlib-metadata==2.1.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
uWSGI==2.0.20
zipp==3.4.1
